rework help toggle to fix parameterized credential page and fix test errors
