+++ app/components/TopBar/TopBar.d.ts
@@ -5,6 +5,4 @@ export type TopBarProps = {
   hasLogo?: boolean;
   setIsDark: function;
   setCredential?: function;
-  isHelpEnabled: boolean;
-  setIsHelpEnabled: function;
 }
