+++ app/layout.tsx
@@ -1,7 +1,10 @@
+"use client"
 import type { Metadata } from "next";
 import { Geist, Geist_Mono } from "next/font/google";
 import "./globals.css";
 import 'material-icons/iconfont/material-icons.css';
+import { useEffect, useState } from "react";
+import { HelpContext } from '@/lib/HelpContext'
 
 const geistSans = Geist({
   variable: "--font-geist-sans",
@@ -13,7 +16,7 @@ const geistMono = Geist_Mono({
   subsets: ["latin"],
 });
 
-export const metadata: Metadata = {
+const metadata: Metadata = {
   title: "VerifierPlus",
   description: "Academic Credential Verification",
 };
@@ -23,11 +26,35 @@ export default function RootLayout({
 }: Readonly<{
   children: React.ReactNode;
 }>) {
+
+  const [isHelpEnabled, setIsHelpEnabled] = useState(true);
+
+  const toggleHelp = () => {
+    if (isHelpEnabled) {
+      setIsHelpEnabled(false);
+      localStorage.setItem('helpMode', 'false');
+    } else {
+      setIsHelpEnabled(true);
+      localStorage.setItem('helpMode', 'true');
+    }   
+  }
+   // get local storage value for help mode on mount
+  useEffect(() => {
+    let helpMode = localStorage.getItem('helpMode');
+    if (helpMode === 'true') { 
+      setIsHelpEnabled(true); 
+    } else { 
+      setIsHelpEnabled(false); 
+    }
+  },[]);
+  
   return (
     <html lang="en">
       <body className={`${geistSans.variable} ${geistMono.variable} antialiased`}>
         <div className="root">
-          {children}
+          <HelpContext.Provider value={{isHelpEnabled, toggleHelp}}>
+            {children}
+          </HelpContext.Provider>
         </div>
       </body>
     </html>
