+++ tests/verifyText.spec.ts
@@ -57,5 +57,5 @@ test('invalid', async ({ page }) => {
     await page.goto("/")
     await page.getByTestId('vc-text-area').fill('blah')
     await page.getByRole('button', { name: 'Verify' }).click()
-    await expect(page.getByText('The JSON is not a Verifiable Credential')).toBeVisible()
+    await expect(page.getByText('Not a Verifiable Credential or an Open Badge 3.0')).toBeVisible()
 })
